refactor: WorldModel Phase 4 — extract state, reasoning, observation, prediction, governance
12 modules extracted (1,776 lines):
- state_save.py, state_load.py, state_status.py, state_metrics.py, state_resources.py
- reasoning_dispatch.py, reasoning_routing.py, reasoning_normalize.py
- observation_update.py, prediction_dispatch.py
- governance.py, governance_causal.py

world_model_core.py: 3,799 → 2,287 lines (-1,512).
Total WorldModel decomposition: 9,583 → 2,287 lines (-76%, 23 modules).

Phases 1-4 of 8 complete. WorldModel decomposition finished.
Remaining: __init__ (582 lines) stays on class for now.
