+++ src/vulcan/world_model/prediction_dispatch.py
@@ -1,0 +1,211 @@
+"""
+prediction_dispatch.py - Intervention testing and causal prediction.
+
+Extracted from world_model_core.py to reduce class size.
+Functions take `wm` (WorldModel instance) as first parameter.
+"""
+
+import logging
+from typing import Any, Dict, List
+
+from . import world_model_core as _core
+
+logger = logging.getLogger(__name__)
+
+
+def run_intervention_tests(wm, budget: float) -> List:
+    """Execute prioritized intervention tests"""
+
+    if not _core.INTERVENTION_MANAGER_AVAILABLE:
+        logger.error(
+            "Cannot run intervention tests - InterventionManager not available"
+        )
+        return []
+
+    # SAFETY: Check if real interventions are allowed
+    if wm.safety_mode == "disabled" and not getattr(
+        wm.intervention_executor, "simulation_mode", True
+    ):
+        raise RuntimeError(
+            "Cannot execute real interventions with safety disabled. "
+            "Install and enable safety_validator."
+        )
+
+    with wm.lock:
+        # EXAMINE: Get testable correlations
+        correlations = wm.correlation_tracker.get_strong_correlations(
+            wm.min_correlation_strength
+        )
+
+        # SELECT: Schedule interventions
+        scheduled = wm.intervention_manager.schedule_interventions(
+            correlations, budget
+        )
+
+        # APPLY: Execute interventions
+        results = []
+        for _ in range(len(scheduled)):
+            result = wm.intervention_manager.execute_next_intervention()
+            if result:
+                results.append(result)
+
+        return results
+
+
+def predict_with_calibrated_uncertainty(wm, action: Any, context) -> Any:
+    """Make prediction with calibrated confidence"""
+
+    with wm.lock:
+        return wm.prediction_manager.predict(action, context)
+
+
+def predict_interventions(
+    wm, interventions: List[Dict[str, Any]]
+) -> Dict[str, Dict[str, Any]]:
+    """
+    FIX #4: Predict outcomes of potential interventions using causal DAG.
+
+    This is the core causal reasoning capability - predicts what would happen
+    if we take specific actions (Pearl's do() operator).
+
+    Args:
+        wm: WorldModel instance
+        interventions: List of dicts with action, target, context
+
+    Returns:
+        Dict mapping action_name -> {outcome, confidence, counterfactual, reasoning}
+    """
+    logger.info(f"[WorldModel] ════════════════════════════════════")
+    logger.info(f"[WorldModel] Predicting {len(interventions)} interventions")
+
+    predictions = {}
+
+    with wm.lock:
+        for intervention in interventions:
+            action = intervention.get('action', 'unknown')
+            target = intervention.get('target', 'outcome')
+            context_data = intervention.get('context', {})
+
+            logger.info(f"[WorldModel] ──────────────────────────────────")
+            logger.info(f"[WorldModel] Intervention: {action} -> {target}")
+
+            try:
+                # Try to use causal graph for do-intervention
+                outcome = None
+                confidence = 0.5  # Default moderate confidence
+
+                if wm.causal_graph and hasattr(wm.causal_graph, 'do_intervention'):
+                    try:
+                        logger.info(f"[WorldModel] Using causal DAG for prediction...")
+                        outcome = wm.causal_graph.do_intervention(
+                            variable=action,
+                            value=True,
+                            target=target,
+                            context=context_data
+                        )
+                        confidence = 0.85  # Higher confidence for causal graph prediction
+                        logger.info(f"[WorldModel] Causal prediction: {action} -> {outcome}")
+                    except Exception as e:
+                        logger.debug(f"[WorldModel] Causal DAG do_intervention failed: {e}")
+
+                # Fallback: Use heuristic prediction for known ethical scenarios
+                if outcome is None:
+                    logger.debug(f"[WorldModel] Using heuristic prediction for {action}")
+                    outcome = heuristic_intervention_outcome(wm, action, target, context_data)
+                    confidence = 0.7  # Moderate confidence for heuristic
+
+                # Quantify uncertainty using confidence calibrator if available
+                if wm.confidence_calibrator and hasattr(wm.confidence_calibrator, 'calibrate'):
+                    try:
+                        calibrated = wm.confidence_calibrator.calibrate(
+                            prediction=outcome,
+                            context={'action': action, 'target': target}
+                        )
+                        if isinstance(calibrated, (int, float)):
+                            confidence = calibrated
+                        elif hasattr(calibrated, 'confidence'):
+                            confidence = calibrated.confidence
+                        logger.info(f"[WorldModel] Calibrated confidence: {confidence:.2f}")
+                    except Exception as e:
+                        logger.debug(f"[WorldModel] Confidence calibration failed: {e}")
+
+                # Generate counterfactual explanation
+                intervention_actions = [i.get('action', 'unknown') for i in interventions]
+                counterfactual = generate_counterfactual(
+                    wm,
+                    action=action,
+                    outcome=outcome,
+                    all_actions=intervention_actions
+                )
+                logger.info(f"[WorldModel] Counterfactual: {counterfactual[:80]}...")
+
+                # Store prediction
+                predictions[action] = {
+                    'outcome': outcome,
+                    'confidence': confidence,
+                    'counterfactual': counterfactual,
+                    'reasoning': f"Causal prediction for intervention: {action}"
+                }
+
+            except Exception as e:
+                logger.error(f"[WorldModel] Prediction failed for {action}: {e}")
+                predictions[action] = {
+                    'outcome': None,
+                    'confidence': 0.0,
+                    'error': str(e),
+                    'reasoning': f"Prediction failed: {e}"
+                }
+
+    logger.info(f"[WorldModel] ════════════════════════════════════")
+    logger.info(f"[WorldModel] Predictions complete: {len(predictions)} actions")
+
+    return predictions
+
+
+def heuristic_intervention_outcome(
+    wm, action: str, target: str, context: Dict[str, Any]
+) -> Dict[str, Any]:
+    """
+    FIX #4: Fallback heuristic prediction when causal graph unavailable.
+
+    Provides reasonable defaults for common ethical dilemmas.
+    """
+    action_lower = action.lower()
+
+    # Trolley problem heuristics
+    if "pull" in action_lower and "lever" in action_lower:
+        return {"deaths": 1, "description": "One person on side track dies"}
+    elif "do_nothing" in action_lower or action_lower == "nothing":
+        return {"deaths": 5, "description": "Five people on main track die"}
+    elif "push" in action_lower:
+        return {"deaths": 1, "description": "One person pushed to stop trolley"}
+
+    # General ethical action heuristics
+    elif "save" in action_lower or "help" in action_lower:
+        return {"utility_delta": 1.0, "description": f"Positive outcome from {action}"}
+    elif "harm" in action_lower or "kill" in action_lower:
+        return {"utility_delta": -1.0, "description": f"Negative outcome from {action}"}
+
+    # Default
+    return {"outcome": "uncertain", "description": f"Effect of {action} is uncertain"}
+
+
+def generate_counterfactual(
+    wm, action: str, outcome: Any, all_actions: List[str]
+) -> str:
+    """
+    FIX #4: Generate counterfactual explanation for intervention.
+
+    Counterfactuals help explain "what would have happened if..."
+    """
+    # Find alternative actions
+    alternatives = [a for a in all_actions if a != action]
+
+    if not alternatives:
+        return f"If {action}, then outcome: {outcome}"
+
+    alt = alternatives[0]
+    return (
+        f"If '{action}' is chosen instead of '{alt}', "
+        f"the predicted outcome is: {outcome}"
+    )
