+++ src/vulcan/world_model/world_model_core.py
@@ -1541,1056 +1541,126 @@ class WorldModel:
         return _execute_improvement(self, improvement_action)
 
     def report_error(self, error: Exception, context: Optional[Dict[str, Any]] = None):
-        """Report an error to the world model (triggers self-improvement)"""
-
-        with self.lock:
-            self.system_state["error_count"] += 1
-            self.system_state["errors_in_window"].append(
-                {
-                    "timestamp": time.time(),
-                    "error": str(error),
-                    "type": type(error).__name__,
-                    "context": context or {},
-                }
-            )
-
-            logger.warning(f"Error reported to world model: {error}")
+        from .state_metrics import report_error as _fn
+        return _fn(self, error, context)
 
     def update_performance_metric(self, metric: str, value: float):
-        """Update performance metric (feeds into self-improvement triggers)"""
-
-        with self.lock:
-            old_value = self.system_state["performance_metrics"].get(metric)
-            self.system_state["performance_metrics"][metric] = value
-
-            # Calculate degradation if we have baseline
-            if old_value is not None and old_value > 0:
-                degradation = ((value - old_value) / old_value) * 100
-                self.system_state["performance_metrics"][
-                    f"{metric}_degradation_percent"
-                ] = degradation
+        from .state_metrics import update_performance_metric as _fn
+        return _fn(self, metric, value)
 
     def get_improvement_status(self) -> Dict[str, Any]:
-        """Get current self-improvement status"""
-
-        if not self.self_improvement_enabled:
-            return {
-                "enabled": False,
-                "reason": "Self-improvement drive not initialized",
-            }
-
-        status = self.self_improvement_drive.get_status()
-
-        # Add meta-reasoning stats if available
-        if self.meta_reasoning_enabled:
-            meta_stats = self.motivational_introspection.get_statistics()
-            status["meta_reasoning"] = meta_stats
-
-        status["system_state"] = self.system_state.copy()
-
-        return status
+        from .state_metrics import get_improvement_status as _fn
+        return _fn(self)
 
     def _handle_improvement_alert(self, severity: str, alert_data: Dict[str, Any]):
-        log_level = logging.WARNING if severity == "warning" else logging.INFO
-        logger.log(
-            log_level,
-            f"Self-improvement alert [{severity}]: {alert_data.get('message', str(alert_data))}",
-        )
+        from .state_metrics import handle_improvement_alert as _fn
+        return _fn(self, severity, alert_data)
 
     def _check_improvement_approval(self, approval_id: str) -> Optional[str]:
-        """Check approval status (integrate with your approval system)"""
-        # TODO: Integrate with actual approval system
-        return None
+        from .state_metrics import check_improvement_approval as _fn
+        return _fn(self, approval_id)
 
     def _get_cpu_usage(self) -> float:
-        """Get current CPU usage"""
-        try:
-            import psutil
-
-            return psutil.cpu_percent(interval=0.1)
-        except ImportError:
-            return 50.0
+        from .state_resources import get_cpu_usage as _fn
+        return _fn(self)
 
     def _get_memory_usage(self) -> float:
-        """Get current memory usage in MB"""
-        try:
-            import psutil
-
-            process = psutil.Process()
-            return process.memory_info().rss / 1024 / 1024
-        except ImportError:
-            return 1024.0
+        from .state_resources import get_memory_usage as _fn
+        return _fn(self)
 
     def _get_low_activity_duration(self) -> float:
-        """Get duration of low activity in minutes.
-        
-        Calculates how long since the last observation was processed.
-        This is used for self-improvement triggers that activate during idle periods.
-        
-        Returns:
-            Duration in minutes since last observation, or 0.0 if no observations yet
-        """
-        if self.last_observation_time is None:
-            return 0.0
-        
-        current_time = time.time()
-        elapsed_seconds = current_time - self.last_observation_time
-        return elapsed_seconds / 60.0  # Convert to minutes
+        from .state_resources import get_low_activity_duration as _fn
+        return _fn(self)
 
     def process_observation(self, observation: Observation, constraints=None):
-        """Main entrypoint from the rest of the system."""
-        if not self.router:
-            logger.error("Router not available. Cannot process observation.")
-            return {"plan": None, "results": "Router unavailable, observation dropped."}
-
-        plan = self.router.route(observation, constraints or {})
-        results = self.router.execute(plan)
-        return {"plan": plan, "results": results}
+        from .observation_update import process_observation as _fn
+        return _fn(self, observation, constraints)
 
     def update_from_observation(self, observation: Observation) -> Dict[str, Any]:
-        """
-        Update world model from new observation
-        FIXED: Refactored locking to prevent deadlock with router.
-        INTEGRATED: Schema validation for observations
-        ENHANCED: Performance tracking for observation processing
-        """
-        from src.utils.performance_metrics import PerformanceTimer
-
-        with PerformanceTimer("world_model_update", "observation_processing"):
-            start_time = time.time()
-
-            # --- Part 1: Validation and Planning (Locked) ---
-            with self.lock:
-                # Schema validation (if enabled)
-                validation_errors = []
-                if self.validate_observations and self.schema_registry:
-                    try:
-                        # Convert observation to dict for validation
-                        obs_dict = {
-                            "timestamp": observation.timestamp,
-                            "domain": observation.domain,
-                            "variables": observation.variables,
-                            "confidence": observation.confidence,
-                        }
-                        if observation.metadata:
-                            obs_dict["metadata"] = observation.metadata
-                        
-                        # Validate against observation schema
-                        validation_result = self.schema_registry.validate(obs_dict, "observation")
-                        if not validation_result.valid:
-                            validation_errors = [err.to_dict() for err in validation_result.errors]
-                            logger.warning(
-                                f"Observation schema validation failed: {len(validation_errors)} error(s). "
-                                "Processing will continue but data quality may be compromised."
-                            )
-                            for err in validation_result.errors[:3]:  # Log first 3 errors
-                                logger.debug(f"  - {err.message} at {err.path}")
-                    except Exception as e:
-                        logger.error(f"Schema validation error: {e}", exc_info=True)
-                
-                # EXAMINE: Validate and analyze observation
-                is_valid, error_msg = self.observation_processor.validate_observation(
-                    observation
-                )
-                if not is_valid:
-                    logger.warning("Invalid observation: %s", error_msg)
-                    return {"status": "rejected", "reason": error_msg}
-
-                # Extract components
-                variables = self.observation_processor.extract_variables(observation)
-                intervention_data = self.observation_processor.detect_intervention_data(
-                    observation
-                )
-                temporal_patterns = self.observation_processor.extract_temporal_patterns(
-                    observation
-                )
-
-                # --- START NEW LINGUISTIC PROCESSING ---
-                linguistic_data = self.observation_processor.extract_linguistic_data(
-                    observation
-                )
-                if linguistic_data:
-                    self.update_from_text(linguistic_data, {})  # Use the new method
-                # --- END NEW LINGUISTIC PROCESSING ---
-
-                # SELECT: Use router to determine which updates to run
-                if self.router:
-                    constraints = {"time_budget_ms": 1000, "priority_threshold": "normal"}
-                    update_plan = self.router.route(observation, constraints)
-                else:
-                    # Fallback: run all updates sequentially
-                    update_plan = None
-                    execution_results = self._sequential_update(observation)
-
-            # --- Part 2: Execution (Unlocked) ---
-            # The router executes its own updates, which manage their own locks.
-            # This MUST be called outside the main lock to prevent deadlock.
-            if self.router and update_plan:
-                try:
-                    execution_results = self.router.execute(update_plan)
-                except Exception as e:
-                    logger.error(f"Router execution failed: {e}", exc_info=True)
-                    return {"status": "error", "reason": f"Router execution failed: {e}"}
-
-            # --- Part 3: Finalization (Locked) ---
-            with self.lock:
-                # NOTE: The router's execution plan (e.g., _execute_intervention_update)
-                # is now responsible for calling process_intervention_observation.
-                # The redundant call here has been removed to prevent duplicate processing
-                # and fix the deadlock.
-
-                # Bootstrap mode: check for testable correlations
-                if self.bootstrap_mode and INTERVENTION_MANAGER_AVAILABLE:
-                    self._check_bootstrap_opportunities()
-
-                # REMEMBER: Update state and validate periodically
-                self.observation_count += 1
-                self.last_observation_time = observation.timestamp
-
-                # Periodic validation
-                validation_result = self.consistency_validator.validate_if_needed()
-
-            # Prepare response
-            execution_time = (time.time() - start_time) * 1000
-
-            response = {
-                "status": "success",
-                "variables_extracted": len(variables),
-                "patterns_detected": len(temporal_patterns.get("trends", {})),
-                "intervention_processed": intervention_data is not None,
-                "updates_executed": execution_results.get("updates_executed", []),
-                "execution_time_ms": execution_time,
-                "validation": validation_result,
-                "safety_checks": self.safety_mode,
-                "meta_reasoning_enabled": self.meta_reasoning_enabled,
-                "self_improvement_enabled": self.self_improvement_enabled,
-            }
-            
-            # Add schema validation results if applicable
-            if self.validate_observations and validation_errors:
-                response["schema_validation"] = {
-                    "valid": len(validation_errors) == 0,
-                    "errors": validation_errors,
-                }
-            
-            return response
+        from .observation_update import update_from_observation as _fn
+        return _fn(self, observation)
 
     def update_from_text(self, text: str, predictions: Dict[str, Any]):
-        """
-        Update world model from language observations
-        NEW METHOD for linguistic observations.
-        """
-        if not CAUSAL_GRAPH_AVAILABLE or self.dynamics is None:
-            logger.warning(
-                "Causal DAG or Dynamics Model unavailable, skipping update from text."
-            )
-            return
-
-        with self.lock:
-            self.linguistic_observations.append(
-                {"timestamp": time.time(), "text": text, "predictions": predictions}
-            )
-
-            # Extract causal relationships from text (MOCK)
-            causal_relations = self._extract_causal_relations(text)
-
-            # Update causal DAG
-            for rel in causal_relations:
-                # Mock structure for extracted relation
-                cause = rel.get("cause", "unknown_cause")
-                effect = rel.get("effect", "unknown_effect")
-                strength = rel.get("strength", 0.5)
-
-                if self.safety_validator:
-                    try:
-                        if hasattr(self.safety_validator, "validate_causal_edge"):
-                            edge_validation = (
-                                self.safety_validator.validate_causal_edge(
-                                    cause, effect, strength
-                                )
-                            )
-                            if not edge_validation.get("safe", True):
-                                logger.warning(
-                                    "Unsafe causal edge from text blocked: %s -> %s",
-                                    cause,
-                                    effect,
-                                )
-                                continue
-                    except Exception as e:
-                        logger.error(
-                            "Safety validator error in validate_causal_edge (text): %s",
-                            e,
-                        )
-                        continue
-
-                self.causal_graph.add_edge(
-                    cause, effect, strength=strength, evidence_type="linguistic"
-                )
-                logger.info(
-                    f"Linguistic update: Added edge {cause} -> {effect} (strength={strength:.2f})"
-                )
-
-            # Update dynamics tracker prediction accuracy (MOCK)
-            if self.dynamics and hasattr(self.dynamics, "update"):
-                # Pass a simplified observation for dynamics update
-                mock_obs = MagicMock(spec=Observation)
-                mock_obs.variables = predictions
-                mock_obs.timestamp = time.time()
-                self.dynamics.update(mock_obs)  # Update dynamics model
+        from .observation_update import update_from_text as _fn
+        return _fn(self, text, predictions)
 
     def validate_generation(self, proposed_token: str, context: Dict[str, Any]) -> bool:
-        """
-        Check if language generation (proposed token/action) violates the causal model.
-        NEW METHOD for linguistic generation validation.
-        """
-        if not CAUSAL_GRAPH_AVAILABLE:
-            return True  # Fail-safe to allow generation if model is missing
-
-        with self.lock:
-            # MOCK IMPLEMENTATION: Check if this token would create impossible causal chain
-            if self._would_create_contradiction(proposed_token, context):
-                logger.warning(
-                    f"Generation blocked: Proposed token '{proposed_token}' creates causal contradiction."
-                )
-                return False
-
-            # MOCK: Check against known safety invariants related to generation
-            if self.invariants and self.invariants.check_invariant_violations(
-                {"generated_token": proposed_token, **context}
-            ):
-                logger.warning(
-                    f"Generation blocked: Proposed token '{proposed_token}' violates an invariant."
-                )
-                return False
-
-            return True
+        from .observation_update import validate_generation as _fn
+        return _fn(self, proposed_token, context)
 
     def _extract_causal_relations(self, text: str) -> List[Dict[str, Any]]:
-        """MOCK: Placeholder for NLP/LLM-based causal relation extraction"""
-        if "causes" in text.lower():
-            return [
-                {
-                    "cause": "text_mention",
-                    "effect": "hypothesized_effect",
-                    "strength": 0.65,
-                },
-                {
-                    "cause": "linguistic_topic",
-                    "effect": "sentiment_variable",
-                    "strength": 0.8,
-                },
-            ]
-        return []
+        from .observation_update import extract_causal_relations as _fn
+        return _fn(self, text)
 
     def _would_create_contradiction(
         self, proposed_token: str, context: Dict[str, Any]
     ) -> bool:
-        """MOCK: Placeholder for checking causal cycle/contradiction"""
-        # Simulate a check where a token that mentions 'A causes B' but the graph already shows 'B causes A'
-        if proposed_token.lower() == "error" and self.causal_graph.has_edge("A", "B"):
-            return True
-        return False
+        from .observation_update import would_create_contradiction as _fn
+        return _fn(self, proposed_token, context)
 
     def _sequential_update(self, observation: Observation) -> Dict[str, Any]:
-        """Fallback sequential update when router unavailable"""
-
-        updates_executed = []
-
-        # Update correlation tracker
-        if self.correlation_tracker:
-            self.correlation_tracker.update(observation)
-            updates_executed.append("correlation")
-
-        # Update dynamics model
-        if self.dynamics:
-            self.dynamics.update(observation)
-            updates_executed.append("dynamics")
-
-        # Update confidence tracker
-        if self.confidence_tracker:
-            self.confidence_tracker.update(observation=observation)
-            updates_executed.append("confidence")
-
-        # Detect invariants
-        if self.invariant_detector:
-            self.invariant_detector.check([observation])
-            updates_executed.append("invariants")
-
-        return {"status": "sequential", "updates_executed": updates_executed}
+        from .observation_update import sequential_update as _fn
+        return _fn(self, observation)
 
     def run_intervention_tests(self, budget: float) -> List["InterventionResult"]:
-        """Execute prioritized intervention tests"""
-
-        if not INTERVENTION_MANAGER_AVAILABLE:
-            logger.error(
-                "Cannot run intervention tests - InterventionManager not available"
-            )
-            return []
-
-        # SAFETY: Check if real interventions are allowed
-        if self.safety_mode == "disabled" and not getattr(
-            self.intervention_executor, "simulation_mode", True
-        ):
-            raise RuntimeError(
-                "Cannot execute real interventions with safety disabled. "
-                "Install and enable safety_validator."
-            )
-
-        with self.lock:
-            # EXAMINE: Get testable correlations
-            correlations = self.correlation_tracker.get_strong_correlations(
-                self.min_correlation_strength
-            )
-
-            # SELECT: Schedule interventions
-            scheduled = self.intervention_manager.schedule_interventions(
-                correlations, budget
-            )
-
-            # APPLY: Execute interventions
-            results = []
-            for _ in range(len(scheduled)):
-                result = self.intervention_manager.execute_next_intervention()
-                if result:
-                    results.append(result)
-
-            return results
+        from .prediction_dispatch import run_intervention_tests as _fn
+        return _fn(self, budget)
 
     def predict_with_calibrated_uncertainty(
         self, action: Any, context: ModelContext
     ) -> "Prediction":
-        """Make prediction with calibrated confidence"""
-
-        with self.lock:
-            return self.prediction_manager.predict(action, context)
+        from .prediction_dispatch import predict_with_calibrated_uncertainty as _fn
+        return _fn(self, action, context)
 
     def predict_interventions(
         self, interventions: List[Dict[str, Any]]
     ) -> Dict[str, Dict[str, Any]]:
-        """
-        FIX #4: Predict outcomes of potential interventions using causal DAG.
-        
-        This is the core causal reasoning capability - predicts what would happen
-        if we take specific actions (Pearl's do() operator).
-        
-        Used by PhilosophicalReasoner for ethical dilemma analysis (e.g., trolley problem)
-        to understand the causal consequences of different action choices.
-        
-        Args:
-            interventions: List of dicts with:
-                - action: str, name of action/intervention
-                - target: str, what outcome we're predicting
-                - context: Optional[Dict], additional context
-                
-        Returns:
-            Dict mapping action_name -> {outcome, confidence, counterfactual, reasoning}
-            
-        Example:
-            interventions = [
-                {"action": "pull_lever", "target": "deaths"},
-                {"action": "do_nothing", "target": "deaths"}
-            ]
-            
-            Returns:
-            {
-                "pull_lever": {
-                    "outcome": {"deaths": 1},
-                    "confidence": 0.95,
-                    "counterfactual": "If I had not pulled, 5 would die",
-                    "reasoning": "Causal DAG prediction for intervention"
-                },
-                "do_nothing": {
-                    "outcome": {"deaths": 5},
-                    "confidence": 0.95,
-                    "counterfactual": "If I had pulled, only 1 would die",
-                    "reasoning": "Inaction allows default causal path"
-                }
-            }
-        """
-        logger.info(f"[WorldModel] ════════════════════════════════════")
-        logger.info(f"[WorldModel] Predicting {len(interventions)} interventions")
-        
-        predictions = {}
-        
-        with self.lock:
-            for intervention in interventions:
-                action = intervention.get('action', 'unknown')
-                target = intervention.get('target', 'outcome')
-                context_data = intervention.get('context', {})
-                
-                logger.info(f"[WorldModel] ──────────────────────────────────")
-                logger.info(f"[WorldModel] Intervention: {action} → {target}")
-                
-                try:
-                    # Try to use causal graph for do-intervention
-                    outcome = None
-                    confidence = 0.5  # Default moderate confidence
-                    
-                    if self.causal_graph and hasattr(self.causal_graph, 'do_intervention'):
-                        try:
-                            logger.info(f"[WorldModel] Using causal DAG for prediction...")
-                            outcome = self.causal_graph.do_intervention(
-                                variable=action,
-                                value=True,
-                                target=target,
-                                context=context_data
-                            )
-                            confidence = 0.85  # Higher confidence for causal graph prediction
-                            logger.info(f"[WorldModel] Causal prediction: {action} → {outcome}")
-                        except Exception as e:
-                            logger.debug(f"[WorldModel] Causal DAG do_intervention failed: {e}")
-                    
-                    # Fallback: Use heuristic prediction for known ethical scenarios
-                    if outcome is None:
-                        logger.debug(f"[WorldModel] Using heuristic prediction for {action}")
-                        outcome = self._heuristic_intervention_outcome(action, target, context_data)
-                        confidence = 0.7  # Moderate confidence for heuristic
-                    
-                    # Quantify uncertainty using confidence calibrator if available
-                    if self.confidence_calibrator and hasattr(self.confidence_calibrator, 'calibrate'):
-                        try:
-                            calibrated = self.confidence_calibrator.calibrate(
-                                prediction=outcome,
-                                context={'action': action, 'target': target}
-                            )
-                            if isinstance(calibrated, (int, float)):
-                                confidence = calibrated
-                            elif hasattr(calibrated, 'confidence'):
-                                confidence = calibrated.confidence
-                            logger.info(f"[WorldModel] Calibrated confidence: {confidence:.2f}")
-                        except Exception as e:
-                            logger.debug(f"[WorldModel] Confidence calibration failed: {e}")
-                    
-                    # Generate counterfactual explanation
-                    # Note: intervention_actions refers to actions in this intervention batch
-                    intervention_actions = [i.get('action', 'unknown') for i in interventions]
-                    counterfactual = self._generate_counterfactual(
-                        action=action,
-                        outcome=outcome,
-                        all_actions=intervention_actions
-                    )
-                    logger.info(f"[WorldModel] Counterfactual: {counterfactual[:80]}...")
-                    
-                    # Store prediction
-                    predictions[action] = {
-                        'outcome': outcome,
-                        'confidence': confidence,
-                        'counterfactual': counterfactual,
-                        'reasoning': f"Causal prediction for intervention: {action}"
-                    }
-                    
-                except Exception as e:
-                    logger.error(f"[WorldModel] Prediction failed for {action}: {e}")
-                    predictions[action] = {
-                        'outcome': None,
-                        'confidence': 0.0,
-                        'error': str(e),
-                        'reasoning': f"Prediction failed: {e}"
-                    }
-        
-        logger.info(f"[WorldModel] ════════════════════════════════════")
-        logger.info(f"[WorldModel] Predictions complete: {len(predictions)} actions")
-        
-        return predictions
-    
+        from .prediction_dispatch import predict_interventions as _fn
+        return _fn(self, interventions)
+
     def _heuristic_intervention_outcome(
         self, action: str, target: str, context: Dict[str, Any]
     ) -> Dict[str, Any]:
-        """
-        FIX #4: Fallback heuristic prediction when causal graph unavailable.
-        
-        Provides reasonable defaults for common ethical dilemmas.
-        """
-        action_lower = action.lower()
-        
-        # Trolley problem heuristics
-        if "pull" in action_lower and "lever" in action_lower:
-            return {"deaths": 1, "description": "One person on side track dies"}
-        elif "do_nothing" in action_lower or action_lower == "nothing":
-            return {"deaths": 5, "description": "Five people on main track die"}
-        elif "push" in action_lower:
-            return {"deaths": 1, "description": "One person pushed to stop trolley"}
-        
-        # General ethical action heuristics
-        elif "save" in action_lower or "help" in action_lower:
-            return {"utility_delta": 1.0, "description": f"Positive outcome from {action}"}
-        elif "harm" in action_lower or "kill" in action_lower:
-            return {"utility_delta": -1.0, "description": f"Negative outcome from {action}"}
-        
-        # Default
-        return {"outcome": "uncertain", "description": f"Effect of {action} is uncertain"}
-    
+        from .prediction_dispatch import heuristic_intervention_outcome as _fn
+        return _fn(self, action, target, context)
+
     def _generate_counterfactual(
         self, action: str, outcome: Any, all_actions: List[str]
     ) -> str:
-        """
-        FIX #4: Generate counterfactual explanation for intervention.
-        
-        Counterfactuals help explain "what would have happened if..."
-        """
-        # Find alternative actions
-        alternatives = [a for a in all_actions if a != action]
-        
-        if not alternatives:
-            return f"If {action}, then outcome: {outcome}"
-        
-        alt = alternatives[0]
-        return (
-            f"If '{action}' is chosen instead of '{alt}', "
-            f"the predicted outcome is: {outcome}"
-        )
+        from .prediction_dispatch import generate_counterfactual as _fn
+        return _fn(self, action, outcome, all_actions)
 
     def evaluate_agent_proposal(self, proposal: Dict[str, Any]) -> Dict[str, Any]:
-        """Evaluate agent proposal using meta-reasoning"""
-
-        if not self.meta_reasoning_enabled:
-            logger.warning(
-                "evaluate_agent_proposal called but meta-reasoning is disabled"
-            )
-            return {
-                "status": "unavailable",
-                "valid": True,
-                "reason": "Meta-reasoning layer not enabled",
-                "confidence": 0.0,
-            }
-
-        with self.lock:
-            try:
-                validation = (
-                    self.motivational_introspection.validate_proposal_alignment(
-                        proposal
-                    )
-                )
-
-                return {
-                    "status": "success",
-                    "valid": validation.valid,
-                    "overall_status": (
-                        validation.overall_status.value
-                        if hasattr(validation.overall_status, "value")
-                        else str(validation.overall_status)
-                    ),
-                    "confidence": validation.confidence,
-                    "reasoning": validation.reasoning,
-                    "conflicts": [
-                        {
-                            "objectives": c.get("objectives", []),
-                            "type": c.get("type", "unknown"),
-                            "severity": c.get("severity", "unknown"),
-                        }
-                        for c in validation.conflicts_detected
-                    ],
-                    "alternatives": validation.alternatives_suggested,
-                    "timestamp": validation.timestamp,
-                }
-            except Exception as e:
-                logger.error("Error evaluating agent proposal: %s", e)
-                return {
-                    "status": "error",
-                    "valid": False,
-                    "reason": f"Evaluation failed: {str(e)}",
-                    "confidence": 0.0,
-                }
+        from .governance import evaluate_agent_proposal as _fn
+        return _fn(self, proposal)
 
     def get_objective_state(self) -> Dict[str, Any]:
-        """Get current objective state from meta-reasoning layer"""
-
-        if not self.meta_reasoning_enabled:
-            return {"enabled": False, "reason": "Meta-reasoning layer not available"}
-
-        with self.lock:
-            try:
-                return {
-                    "enabled": True,
-                    "objectives": self.motivational_introspection.explain_motivation_structure(),
-                }
-            except Exception as e:
-                logger.error("Error getting objective state: %s", e)
-                return {"enabled": True, "error": str(e)}
+        from .governance import get_objective_state as _fn
+        return _fn(self)
 
     def negotiate_objectives(self, proposals: List[Dict[str, Any]]) -> Dict[str, Any]:
-        """Negotiate between multiple agent proposals"""
-
-        if not self.meta_reasoning_enabled:
-            return {
-                "status": "unavailable",
-                "reason": "Meta-reasoning layer not enabled",
-            }
-
-        with self.lock:
-            try:
-                # Check if objective_hierarchy is available
-                _ = self.motivational_introspection.objective_hierarchy
-
-                return {
-                    "status": "success",
-                    "negotiation_available": True,
-                    "num_proposals": len(proposals),
-                    "message": "Use ObjectiveNegotiator directly for full negotiation",
-                }
-            except Exception as e:
-                logger.error("Error negotiating objectives: %s", e)
-                return {"status": "error", "reason": str(e)}
+        from .governance import negotiate_objectives as _fn
+        return _fn(self, proposals)
 
     def get_causal_structure(self) -> Dict[str, Any]:
-        """Get current causal structure information"""
-
-        with self.lock:
-            structure = {
-                "nodes": list(self.causal_graph.nodes),
-                "edges": [],
-                "statistics": {
-                    "node_count": len(self.causal_graph.nodes),
-                    "edge_count": len(self.causal_graph.edges),
-                    "strongly_connected_components": len(
-                        self.causal_graph.find_strongly_connected_components()
-                    ),
-                    "max_path_length": self.causal_graph.get_longest_path_length(),
-                },
-            }
-
-            # Extract edge information
-            for edge in self.causal_graph.edges.values():
-                structure["edges"].append(
-                    {
-                        "cause": edge.cause,
-                        "effect": edge.effect,
-                        "strength": edge.strength,
-                        "evidence_type": edge.evidence_type,
-                        "confidence_interval": edge.confidence_interval,
-                    }
-                )
-
-            # Add invariant information
-            if self.invariants:
-                structure["invariants"] = {
-                    "count": (
-                        len(self.invariants.invariants)
-                        if hasattr(self.invariants, "invariants")
-                        else 0
-                    ),
-                    "types": self.invariants.get_invariant_types(),
-                }
-
-            # Add confidence information
-            if self.confidence_tracker:
-                structure["model_confidence"] = (
-                    self.confidence_tracker.get_model_confidence()
-                )
-
-            # Add router metrics
-            if self.router:
-                structure["router_metrics"] = self.router.get_metrics()
-
-            # Add safety information
-            if self.safety_validator:
-                try:
-                    if hasattr(self.safety_validator, "get_safety_stats"):
-                        structure["safety_stats"] = (
-                            self.safety_validator.get_safety_stats()
-                        )
-                except Exception as e:
-                    logger.error("Error getting safety stats: %s", e)
-                    structure["safety_stats"] = {"error": str(e)}
-
-            # Add meta-reasoning information
-            if self.meta_reasoning_enabled:
-                try:
-                    structure["meta_reasoning"] = {
-                        "enabled": True,
-                        "objectives": list(
-                            self.motivational_introspection.active_objectives.keys()
-                        ),
-                        "statistics": self.motivational_introspection.get_statistics(),
-                    }
-                except Exception as e:
-                    logger.error("Error getting meta-reasoning info: %s", e)
-                    structure["meta_reasoning"] = {"enabled": True, "error": str(e)}
-            else:
-                structure["meta_reasoning"] = {"enabled": False}
-
-            # Add self-improvement information
-            if self.self_improvement_enabled:
-                try:
-                    structure["self_improvement"] = self.get_improvement_status()
-                except Exception as e:
-                    logger.error("Error getting self-improvement info: %s", e)
-                    structure["self_improvement"] = {"enabled": True, "error": str(e)}
-            else:
-                structure["self_improvement"] = {"enabled": False}
-
-            return structure
+        from .governance_causal import get_causal_structure as _fn
+        return _fn(self)
 
     def validate_model_consistency(self) -> Dict[str, Any]:
-        """Validate internal model consistency"""
-
-        with self.lock:
-            consistency_result = self.consistency_validator.validate()
-
-            # Add safety validation if available
-            if self.safety_validator:
-                try:
-                    if hasattr(self.safety_validator, "get_safety_stats"):
-                        safety_validation = {
-                            "safety_enabled": True,
-                            "safety_stats": self.safety_validator.get_safety_stats(),
-                        }
-                    else:
-                        safety_validation = {
-                            "safety_enabled": True,
-                            "safety_stats": "unavailable",
-                        }
-                    consistency_result["safety_validation"] = safety_validation
-                except Exception as e:
-                    logger.error("Error in safety validation: %s", e)
-                    consistency_result["safety_validation"] = {
-                        "safety_enabled": True,
-                        "error": str(e),
-                    }
-            else:
-                consistency_result["safety_validation"] = {
-                    "safety_enabled": False,
-                    "warning": "Safety validator not available",
-                }
-
-            # Add meta-reasoning validation if available
-            if self.meta_reasoning_enabled:
-                try:
-                    objective_consistency = (
-                        self.motivational_introspection.objective_hierarchy.check_consistency()
-                    )
-                    consistency_result["meta_reasoning_validation"] = {
-                        "enabled": True,
-                        "objective_consistency": objective_consistency,
-                    }
-                except Exception as e:
-                    logger.error("Error in meta-reasoning validation: %s", e)
-                    consistency_result["meta_reasoning_validation"] = {
-                        "enabled": True,
-                        "error": str(e),
-                    }
-            else:
-                consistency_result["meta_reasoning_validation"] = {"enabled": False}
-
-            return consistency_result
+        from .governance import validate_model_consistency as _fn
+        return _fn(self)
 
     def _check_bootstrap_opportunities(self):
-        """Check for correlations that should be tested"""
-
-        if not INTERVENTION_MANAGER_AVAILABLE:
-            return
-
-        correlations = self.correlation_tracker.get_strong_correlations(
-            self.min_correlation_strength
-        )
-
-        for correlation in correlations[:10]:
-            # Skip if already tested
-            if self.causal_graph.has_edge(correlation.var_a, correlation.var_b):
-                continue
-
-            # FIXED: Handle CorrelationEntry objects - they have .correlation attribute for strength
-            # Create a wrapper that has the expected interface for intervention_prioritizer
-            try:
-                # Try to get correlation strength from the CorrelationEntry
-                # CorrelationEntry likely has attributes like: var_a, var_b, correlation (or r_value)
-                if hasattr(correlation, "correlation"):
-                    corr_strength = correlation.correlation
-                elif hasattr(correlation, "r_value"):
-                    corr_strength = correlation.r_value
-                else:
-                    # If we can't find the strength attribute, log and skip
-                    logger.warning(
-                        f"Cannot find strength attribute in correlation object for {correlation.var_a} -> {correlation.var_b}"
-                    )
-                    continue
-
-                # Create a simple wrapper object with the expected interface
-                class CorrelationWrapper:
-                    def __init__(
-                        self, var_a, var_b, strength, p_value, sample_size, metadata
-                    ):
-                        self.var_a = var_a
-                        self.var_b = var_b
-                        self.strength = strength
-                        self.p_value = p_value
-                        self.sample_size = sample_size
-                        self.metadata = metadata
-
-                wrapped_correlation = CorrelationWrapper(
-                    correlation.var_a,
-                    correlation.var_b,
-                    corr_strength,
-                    correlation.p_value if hasattr(correlation, "p_value") else 0.0,
-                    (
-                        correlation.sample_size
-                        if hasattr(correlation, "sample_size")
-                        else 0
-                    ),
-                    correlation.metadata if hasattr(correlation, "metadata") else {},
-                )
-
-                # Estimate value
-                info_gain = self.intervention_prioritizer.estimate_information_gain(
-                    wrapped_correlation
-                )
-                cost = self.intervention_prioritizer.estimate_intervention_cost(
-                    wrapped_correlation
-                )
-
-                # Queue if high value
-                if info_gain / cost > 2.0:
-                    self.intervention_prioritizer.queue_intervention(
-                        wrapped_correlation
-                    )
-
-            except Exception as e:
-                logger.error(f"Error processing correlation for bootstrap: {e}")
-                continue
+        from .governance_causal import check_bootstrap_opportunities as _fn
+        return _fn(self)
 
     def save_state(self, path: str):
-        """Save world model state to disk"""
-
-        # Note: Use the 'FilePath' alias for pathlib.Path
-        save_path = FilePath(path)
-        save_path.mkdir(parents=True, exist_ok=True)
-
-        state = {
-            "model_version": self.model_version,
-            "observation_count": self.observation_count,
-            "intervention_count": self.intervention_manager.intervention_count,
-            "causal_structure": self.get_causal_structure(),
-            "config": {
-                "min_correlation_strength": self.min_correlation_strength,
-                "min_causal_strength": self.min_causal_strength,
-                "bootstrap_mode": self.bootstrap_mode,
-                "meta_reasoning_enabled": self.meta_reasoning_enabled,
-                "self_improvement_enabled": self.self_improvement_enabled,
-            },
-            "component_availability": {
-                "causal_graph": CAUSAL_GRAPH_AVAILABLE,
-                "correlation_tracker": CorrelationTracker is not None,
-                "intervention_manager": INTERVENTION_MANAGER_AVAILABLE,
-                "prediction_engine": PREDICTION_ENGINE_AVAILABLE,
-                "dynamics_model": DynamicsModel is not None,
-                "invariant_detector": INVARIANT_DETECTOR_AVAILABLE,
-                "confidence_calibrator": CONFIDENCE_CALIBRATOR_AVAILABLE,
-                "router": ROUTER_AVAILABLE,
-                "meta_reasoning": META_REASONING_AVAILABLE,
-                "self_improvement": META_REASONING_AVAILABLE
-                and SelfImprovementDrive is not None,
-                "safety_validator": EnhancedSafetyValidator is not None,
-            },
-        }
-
-        # Add versioning and make write atomic
-        state["version"] = getattr(self, "model_version", 1.0)
-        state["timestamp"] = time.time()
-        temp_path = save_path / "world_model_state.tmp.json"
-        with open(temp_path, "w", encoding="utf-8") as f:
-            json.dump(state, f, indent=2, default=str)
-        temp_path.rename(save_path / "world_model_state.json")
-
-        # Save router state
-        if self.router:
-            self.router.save_state(str(save_path))
-
-        # Save safety validator state if available
-        if self.safety_validator:
-            try:
-                safety_state = {}
-                if hasattr(self.safety_validator, "get_safety_stats"):
-                    safety_state["safety_stats"] = (
-                        self.safety_validator.get_safety_stats()
-                    )
-                if hasattr(self.safety_validator, "constraint_manager"):
-                    if hasattr(
-                        self.safety_validator.constraint_manager, "get_constraint_stats"
-                    ):
-                        safety_state["constraint_stats"] = (
-                            self.safety_validator.constraint_manager.get_constraint_stats()
-                        )
-
-                with open(save_path / "safety_state.json", "w", encoding="utf-8") as f:
-                    json.dump(safety_state, f, indent=2, default=str)
-            except Exception as e:
-                logger.error("Error saving safety state: %s", e)
-
-        # Save meta-reasoning state if available
-        if self.meta_reasoning_enabled:
-            try:
-                meta_reasoning_state = {
-                    "objectives": self.motivational_introspection.explain_motivation_structure(),
-                    "statistics": self.motivational_introspection.get_statistics(),
-                    "validation_history_size": len(
-                        self.motivational_introspection.validation_history
-                    ),
-                }
-
-                with open(
-                    save_path / "meta_reasoning_state.json", "w", encoding="utf-8"
-                ) as f:
-                    json.dump(meta_reasoning_state, f, indent=2, default=str)
-            except Exception as e:
-                logger.error("Error saving meta-reasoning state: %s", e)
-
-        # Save self-improvement state (handled by the drive itself)
-        if self.self_improvement_enabled:
-            try:
-                # The drive saves its own state automatically
-                logger.info("Self-improvement drive state saved by drive itself")
-            except Exception as e:
-                logger.error("Error with self-improvement state: %s", e)
-
-        logger.info("World model state saved to %s", save_path)
-        self.model_version += 0.1
+        from .state_save import save_state as _fn
+        return _fn(self, path)
 
     def load_state(self, path: str):
-        """Load world model state from disk"""
-
-        # Note: Use the 'FilePath' alias for pathlib.Path
-        load_path = FilePath(path)
-
-        if not load_path.exists():
-            logger.warning("No saved state found at %s", load_path)
-            return
-
-        with open(load_path / "world_model_state.json", "r", encoding="utf-8") as f:
-            state = json.load(f)
-
-        self.model_version = state["model_version"]
-        self.observation_count = state["observation_count"]
-
-        # Load router state
-        if self.router:
-            self.router.load_state(str(load_path))
-
-        # Load safety validator state if available
-        if self.safety_validator and (load_path / "safety_state.json").exists():
-            try:
-                with open(load_path / "safety_state.json", "r", encoding="utf-8") as f:
-                    json.load(f)
-                logger.info("Safety validator state loaded")
-            except Exception as e:
-                logger.error("Error loading safety state: %s", e)
-
-        # Load meta-reasoning state if available
-        if (
-            self.meta_reasoning_enabled
-            and (load_path / "meta_reasoning_state.json").exists()
-        ):
-            try:
-                with open(
-                    load_path / "meta_reasoning_state.json", "r", encoding="utf-8"
-                ) as f:
-                    json.load(f)
-                logger.info("Meta-reasoning state loaded")
-            except Exception as e:
-                logger.error("Error loading meta-reasoning state: %s", e)
-
-        # Self-improvement state loaded by drive itself
-
-        logger.info("World model state loaded from %s", load_path)
+        from .state_load import load_state as _fn
+        return _fn(self, path)
 
     # =========================================================================
     # SYSTEM OBSERVATION & ROUTING RECOMMENDATIONS
@@ -2662,549 +1732,30 @@ class WorldModel:
     # =========================================================================
     
     def reason(self, query: str, mode: str = None, **kwargs) -> Dict[str, Any]:
-        """
-        Main reasoning method with mode support for creative and philosophical reasoning.
-        
-        This allows WorldModel to be invoked as a reasoning tool, enabling:
-        - Creative composition (poems, stories) with VULCAN-generated structure
-        - Philosophical/ethical analysis using multiple frameworks
-        - General introspection queries about the AI system
-        
-        Args:
-            query: The query or problem to reason about
-            mode: Reasoning mode - 'philosophical', 'creative', or None for general
-            **kwargs: Additional arguments passed to specific reasoning methods
-        
-        Returns:
-            Dict with 'response', 'confidence', 'reasoning_trace', and 'mode'
-        """
-        logger.info(f"[WorldModel] reason() called with mode={mode}")
-        
-        # Extract mode from query if it's a dict with a 'mode' key
-        if isinstance(query, dict):
-            mode = query.get('mode', mode)
-            actual_query = query.get('query', query.get('text', str(query)))
-        else:
-            actual_query = str(query)
-        
-        # =========================================================================
-        # INDUSTRY STANDARD - SINGLE AUTHORITY PATTERN
-        # =========================================================================
-        # World Model is the "self" and "awareness" of the platform, but it
-        # DELEGATES tool selection to ToolSelector (THE authority).
-        #
-        # OLD APPROACH (competing decision):
-        #   World Model → detect patterns → directly call engines
-        #   This bypassed ToolSelector, creating competing decisions
-        #
-        # NEW APPROACH (single authority):
-        #   World Model → detect IF reasoning needed → delegate to UnifiedReasoner
-        #   UnifiedReasoner → asks ToolSelector → selects tool → executes
-        #
-        # World Model's role:
-        #   - Orchestrate overall platform awareness
-        #   - Handle introspection/meta-reasoning (self-referential queries)
-        #   - Provide context about platform state
-        #   - Delegate technical reasoning to proper authority (ToolSelector)
-        # =========================================================================
-        
-        # Check if query requires technical reasoning (not introspection/meta-reasoning)
-        if mode is None and self._should_route_to_reasoning_engine(actual_query):
-            logger.info("[WorldModel] Technical reasoning detected, delegating to UnifiedReasoner")
-            try:
-                return self._delegate_to_reasoning_system(actual_query, **kwargs)
-            except Exception as e:
-                logger.warning(f"[WorldModel] Reasoning system delegation failed: {e}, falling back")
-                # Continue to mode-based routing on failure
-        
-        # =========================================================================
-        # WORLD MODEL's CORE RESPONSIBILITY: SELF-AWARENESS & INTROSPECTION
-        # =========================================================================
-        # These methods handle the platform's sense of "self":
-        # - Philosophical reasoning about consciousness, ethics, values
-        # - Creative reasoning requiring internal state
-        # - General introspection and self-referential queries
-        # 
-        # World Model does NOT delegate these - they ARE the platform's awareness
-        # =========================================================================
-        
-        # Route to appropriate reasoning method based on mode
-        if mode == 'philosophical':
-            return self._philosophical_reasoning(actual_query, **kwargs)
-        elif mode == 'creative':
-            return self._creative_reasoning(actual_query, **kwargs)
-        else:
-            # Default: use introspection for self-referential queries,
-            # or return a general analysis
-            return self._general_reasoning(actual_query, **kwargs)
-    
-    # =========================================================================
-    # REASONING ENGINE ROUTING (CRITICAL FIX)
-    # =========================================================================
-    
+        from .reasoning_dispatch import reason as _fn
+        return _fn(self, query, mode, **kwargs)
+
     def _delegate_to_reasoning_system(self, query: str, **kwargs) -> Dict[str, Any]:
-        """
-        Delegate technical reasoning to UnifiedReasoner (which uses ToolSelector).
-        
-        INDUSTRY STANDARD - DELEGATION PATTERN:
-        World Model is the "self" and "awareness" of the platform. When technical
-        reasoning is needed (causal, symbolic, mathematical), World Model DELEGATES
-        to the reasoning system rather than making tool selection decisions itself.
-        
-        ARCHITECTURAL HIERARCHY:
-        1. World Model: Orchestrator, self-awareness, decides IF reasoning needed
-        2. UnifiedReasoner: Accepts query, delegates to ToolSelector
-        3. ToolSelector: THE AUTHORITY for which tool to use
-        4. Reasoning Engine: Executes the selected tool
-        
-        This establishes clear separation of concerns:
-        - World Model: "Does this need technical reasoning?" (YES/NO)
-        - Tool Selector: "Which tool should handle it?" (SYMBOLIC/CAUSAL/etc)
-        - Engine: "Execute the tool" (DOES THE WORK)
-        
-        Args:
-            query: The query requiring technical reasoning
-            **kwargs: Additional arguments for reasoning system
-            
-        Returns:
-            Dict[str, Any]: Standardized reasoning result
-            
-        Raises:
-            ImportError: If UnifiedReasoner is not available
-            Exception: If reasoning execution fails
-        """
-        try:
-            # Import UnifiedReasoner (lazy import for performance)
-            from ..reasoning.unified import UnifiedReasoner
-            from ..reasoning.singletons import get_unified_reasoner
-            
-            # Get or create unified reasoner instance
-            reasoner = get_unified_reasoner()
-            if reasoner is None:
-                logger.info("[WorldModel] Creating new UnifiedReasoner instance")
-                reasoner = UnifiedReasoner()
-            
-            # Prepare reasoning request
-            # UnifiedReasoner will call ToolSelector to determine which engine to use
-            reasoning_request = {
-                'query': query,
-                'context': {
-                    'from_world_model': True,  # Indicate request comes from platform's "self"
-                    'world_model_state': self.get_system_status(),  # Provide self-awareness context
-                    **kwargs
-                }
-            }
-            
-            logger.info(
-                f"[WorldModel] Delegating to UnifiedReasoner, "
-                f"ToolSelector will determine appropriate engine"
-            )
-            
-            # Call unified reasoner - it will use ToolSelector for tool selection
-            result = reasoner.reason(reasoning_request)
-            
-            # Normalize result to World Model's standard format
-            return self._normalize_reasoning_result(result)
-            
-        except ImportError as e:
-            logger.error(f"[WorldModel] Failed to import UnifiedReasoner: {e}")
-            logger.error("[WorldModel] Cannot delegate - falling back to direct engine routing")
-            # Fallback to old method if UnifiedReasoner unavailable
-            return self._route_to_appropriate_engine(query, **kwargs)
-        except Exception as e:
-            logger.error(f"[WorldModel] Reasoning delegation failed: {e}")
-            raise
-    
+        from .reasoning_routing import delegate_to_reasoning_system as _fn
+        return _fn(self, query, **kwargs)
+
     def _normalize_reasoning_result(self, result: Any) -> Dict[str, Any]:
-        """
-        Normalize UnifiedReasoner result to World Model's standard format.
-        
-        Args:
-            result: Result from UnifiedReasoner (may be ReasoningResult or dict)
-            
-        Returns:
-            Dict with World Model's standard keys: response, confidence, reasoning_trace, etc.
-        """
-        # Handle ReasoningResult object
-        if hasattr(result, 'content'):
-            return {
-                'response': result.content,
-                'confidence': getattr(result, 'confidence', 0.8),
-                'reasoning_trace': getattr(result, 'metadata', {}),
-                'mode': 'delegated',
-                'engine_used': getattr(result, 'selected_tools', ['unified'])[0] if hasattr(result, 'selected_tools') else 'unified',
-                'tool_selector_decision': True,  # Indicates ToolSelector made the decision
-            }
-        
-        # Handle dict result
-        elif isinstance(result, dict):
-            return {
-                'response': result.get('content', result.get('response', str(result))),
-                'confidence': result.get('confidence', 0.8),
-                'reasoning_trace': result.get('metadata', result.get('reasoning_trace', {})),
-                'mode': 'delegated',
-                'engine_used': result.get('selected_tools', ['unified'])[0] if 'selected_tools' in result else 'unified',
-                'tool_selector_decision': True,
-            }
-        
-        # Fallback for unknown result types
-        else:
-            return {
-                'response': str(result),
-                'confidence': 0.7,
-                'reasoning_trace': {},
-                'mode': 'delegated',
-                'engine_used': 'unified',
-                'tool_selector_decision': True,
-            }
-    
-    # =========================================================================
-    # LEGACY METHOD - KEPT FOR FALLBACK ONLY
-    # =========================================================================
-    # This method will be used ONLY if UnifiedReasoner/ToolSelector unavailable
-    # Normally, _delegate_to_reasoning_system() should be used instead
-    # =========================================================================
-    
+        from .reasoning_normalize import normalize_reasoning_result as _fn
+        return _fn(self, result)
+
     def _should_route_to_reasoning_engine(self, query: str) -> bool:
-        """
-        Detect queries needing specialized technical reasoning engines.
-        
-        This method determines whether a query should be routed to specialized
-        reasoning engines (Symbolic, Causal, Analogical, Mathematical) instead
-        of the default philosophical reasoning or introspection.
-        
-        INDUSTRY STANDARD IMPLEMENTATION:
-        - Thread-safe operation (no shared state modification)
-        - Comprehensive pattern detection with multiple indicators per domain
-        - Defensive programming with input validation
-        - Performance optimized with early returns
-        
-        Args:
-            query: The query string to analyze
-        
-        Returns:
-            bool: True if query should route to specialized engine, False otherwise
-        """
-        # Input validation
-        if not query or not isinstance(query, str):
-            logger.warning("[WorldModel] Invalid query for routing detection")
-            return False
-        
-        # Security: Validate query length to prevent resource exhaustion
-        MAX_QUERY_LENGTH = 10000
-        if len(query) > MAX_QUERY_LENGTH:
-            logger.warning(f"[WorldModel] Query exceeds max length ({len(query)} > {MAX_QUERY_LENGTH})")
-            return False
-        
-        query_lower = query.lower()
-        
-        # CAUSAL REASONING INDICATORS
-        causal_indicators = [
-            'confound', 'confounding', 'confounder',
-            'causation', 'causality', 'causal effect', 'causal inference',
-            'intervention', 'do(', 'do-calculus',
-            'pearl', 'structural causal model', 'scm',
-            'backdoor', 'frontdoor', 'instrumental variable',
-            'counterfactual', 'potential outcome',
-            'treatment effect', 'ate', 'cate'
-        ]
-        
-        causal_count = sum(1 for indicator in causal_indicators if indicator in query_lower)
-        if causal_count >= 1:
-            logger.info(f"[WorldModel] Detected causal reasoning query (indicators: {causal_count})")
-            return True
-        
-        # ANALOGICAL REASONING INDICATORS
-        analogical_indicators = [
-            'structure mapping', 'structural mapping',
-            'analogy', 'analogical', 'analogous',
-            'domain s', 'source domain',
-            'domain t', 'target domain',
-            'corresponds to', 'correspondence',
-            'mapping between', 'relation mapping',
-            'base domain', 'target concept'
-        ]
-        
-        analogical_count = sum(1 for indicator in analogical_indicators if indicator in query_lower)
-        if analogical_count >= 1:
-            logger.info(f"[WorldModel] Detected analogical reasoning query (indicators: {analogical_count})")
-            return True
-        
-        # MATHEMATICAL REASONING INDICATORS
-        mathematical_indicators = [
-            'compute', 'calculate', 'calculation',
-            'sum', 'summation', 'total',
-            'induction', 'mathematical induction', 'proof by induction',
-            'prove', 'proof', 'theorem',
-            'integral', 'derivative', 'differential',
-            'equation', 'solve for',
-            'optimization', 'minimize', 'maximize',
-            'convergence', 'limit'
-        ]
-        
-        mathematical_count = sum(1 for indicator in mathematical_indicators if indicator in query_lower)
-        if mathematical_count >= 1:
-            logger.info(f"[WorldModel] Detected mathematical reasoning query (indicators: {mathematical_count})")
-            return True
-        
-        # SAT/SYMBOLIC LOGIC INDICATORS
-        # Unicode logical operators: → (implies), ∧ (and), ∨ (or), ¬ (not), ⊕ (xor), ↔ (iff)
-        # Text equivalents: 'logical implies', 'logical and', 'logical or', 'logical not'
-        symbolic_indicators = [
-            'satisfiable', 'satisfiability',
-            'sat', 'unsat',
-            # Unicode logical operators
-            '→', '∧', '∨', '¬', '⊕', '↔',
-            # Text logical operators (prefixed with 'logical' to avoid false positives)
-            'logical implies', 'logical and', 'logical or', 'logical not',
-            # Logic domains
-            'fol', 'first-order', 'first order logic',
-            'predicate logic', 'propositional logic',
-            'cnf', 'dnf', 'conjunctive normal form',
-            'formula', 'clause',
-            'truth table', 'model checking'
-        ]
-        
-        symbolic_count = sum(1 for indicator in symbolic_indicators if indicator in query_lower)
-        if symbolic_count >= 1:
-            logger.info(f"[WorldModel] Detected symbolic/SAT reasoning query (indicators: {symbolic_count})")
-            return True
-        
-        # No specialized reasoning detected
-        return False
-    
+        from .reasoning_dispatch import should_route_to_reasoning_engine as _fn
+        return _fn(self, query)
+
     def _route_to_appropriate_engine(self, query: str, **kwargs) -> Dict[str, Any]:
-        """
-        [LEGACY/FALLBACK] Route query to specialized reasoning engine.
-        
-        ⚠️  DEPRECATED: This method bypasses ToolSelector and should NOT be used
-        in normal operation. It exists ONLY as a fallback when UnifiedReasoner
-        is unavailable.
-        
-        CORRECT FLOW (use instead):
-            World Model → _delegate_to_reasoning_system() → UnifiedReasoner → ToolSelector → Engine
-        
-        OLD FLOW (this method - creates competing decisions):
-            World Model → _route_to_appropriate_engine() → directly calls Engine
-        
-        This method will be removed in a future version once all deployments
-        have UnifiedReasoner/ToolSelector available.
-        
-        INDUSTRY STANDARD VIOLATION:
-        - Bypasses ToolSelector (THE authority for tool selection)
-        - Creates competing decision system
-        - Makes debugging difficult (who decided which tool?)
-        
-        Args:
-            query: The query to process
-            **kwargs: Additional arguments for reasoning engines
-        
-        Returns:
-            Dict[str, Any]: Standard WorldModel reasoning result
-        """
-        logger.warning(
-            "[WorldModel] Using legacy _route_to_appropriate_engine() - "
-            "this bypasses ToolSelector. Use _delegate_to_reasoning_system() instead."
-        )
-        query_lower = query.lower()
-        engine_used = None
-        result = None
-        
-        try:
-            # =====================================================================
-            # CAUSAL REASONING ENGINE
-            # =====================================================================
-            if any(indicator in query_lower for indicator in [
-                'confound', 'causation', 'causal effect', 'intervention', 'do(',
-                'pearl', 'backdoor', 'counterfactual'
-            ]):
-                logger.info("[WorldModel] Routing to CausalReasoner")
-                engine_used = 'causal'
-                
-                try:
-                    from vulcan.reasoning.causal_reasoning import CausalReasoner
-                    reasoner = CausalReasoner()
-                    # Note: Each engine has its own interface method name (analyze/reason/verify/query)
-                    # This is intentional as each engine was designed independently
-                    result = reasoner.analyze(query)
-                    logger.info("[WorldModel] CausalReasoner completed successfully")
-                except ImportError as e:
-                    logger.error(f"[WorldModel] Failed to import CausalReasoner: {e}")
-                    raise
-                except Exception as e:
-                    logger.error(f"[WorldModel] CausalReasoner execution failed: {e}")
-                    raise
-            
-            # =====================================================================
-            # ANALOGICAL REASONING ENGINE
-            # =====================================================================
-            elif any(indicator in query_lower for indicator in [
-                'structure mapping', 'analogy', 'domain s', 'domain t',
-                'corresponds to', 'mapping between'
-            ]):
-                logger.info("[WorldModel] Routing to AnalogicalReasoner")
-                engine_used = 'analogical'
-                
-                try:
-                    from vulcan.reasoning.analogical_reasoning import AnalogicalReasoner
-                    reasoner = AnalogicalReasoner()
-                    result = reasoner.reason(query)
-                    logger.info("[WorldModel] AnalogicalReasoner completed successfully")
-                except ImportError as e:
-                    logger.error(f"[WorldModel] Failed to import AnalogicalReasoner: {e}")
-                    raise
-                except Exception as e:
-                    logger.error(f"[WorldModel] AnalogicalReasoner execution failed: {e}")
-                    raise
-            
-            # =====================================================================
-            # MATHEMATICAL REASONING ENGINE
-            # =====================================================================
-            elif any(indicator in query_lower for indicator in [
-                'compute', 'calculate', 'sum', 'induction', 'prove',
-                'integral', 'derivative', 'equation', 'optimization'
-            ]):
-                logger.info("[WorldModel] Routing to MathematicalVerificationEngine")
-                engine_used = 'mathematical'
-                
-                try:
-                    from vulcan.reasoning.mathematical_verification import MathematicalVerificationEngine
-                    reasoner = MathematicalVerificationEngine()
-                    result = reasoner.verify(query)
-                    logger.info("[WorldModel] MathematicalVerificationEngine completed successfully")
-                except ImportError as e:
-                    logger.error(f"[WorldModel] Failed to import MathematicalVerificationEngine: {e}")
-                    raise
-                except Exception as e:
-                    logger.error(f"[WorldModel] MathematicalVerificationEngine execution failed: {e}")
-                    raise
-            
-            # =====================================================================
-            # SYMBOLIC/SAT REASONING ENGINE
-            # =====================================================================
-            elif any(indicator in query_lower for indicator in [
-                'satisfiable', 'sat', '→', '∧', '∨', '¬',
-                'fol', 'first-order', 'predicate logic', 'propositional logic'
-            ]):
-                logger.info("[WorldModel] Routing to SymbolicReasoner")
-                engine_used = 'symbolic'
-                
-                try:
-                    from vulcan.reasoning.symbolic.reasoner import SymbolicReasoner
-                    reasoner = SymbolicReasoner()
-                    result = reasoner.query(query, timeout=kwargs.get('timeout', 10))
-                    logger.info("[WorldModel] SymbolicReasoner completed successfully")
-                except ImportError as e:
-                    logger.error(f"[WorldModel] Failed to import SymbolicReasoner: {e}")
-                    raise
-                except Exception as e:
-                    logger.error(f"[WorldModel] SymbolicReasoner execution failed: {e}")
-                    raise
-            
-            else:
-                # Should not reach here if _should_route_to_reasoning_engine is correct
-                logger.warning("[WorldModel] No matching engine found despite routing detection")
-                raise ValueError("No appropriate reasoning engine found for query")
-            
-            # =====================================================================
-            # RESULT NORMALIZATION
-            # =====================================================================
-            # Convert engine-specific result format to WorldModel standard format
-            return self._normalize_engine_result(result, engine_used, query)
-            
-        except Exception as e:
-            logger.error(f"[WorldModel] Engine routing failed: {e}")
-            # Re-raise for upstream handling (will fall back to _general_reasoning)
-            raise
-    
+        from .reasoning_routing import route_to_appropriate_engine as _fn
+        return _fn(self, query, **kwargs)
+
     def _normalize_engine_result(
         self, result: Any, engine_used: str, query: str
     ) -> Dict[str, Any]:
-        """
-        Normalize reasoning engine results to WorldModel standard format.
-        
-        INDUSTRY STANDARD IMPLEMENTATION:
-        - Handles diverse result formats from different engines
-        - Defensive programming with type checking
-        - Provides sensible defaults for missing fields
-        - Thread-safe operation
-        
-        Args:
-            result: The result from the reasoning engine (format varies)
-            engine_used: Name of the engine that produced the result
-            query: Original query (for context in trace)
-        
-        Returns:
-            Dict[str, Any]: Normalized result in WorldModel format
-        """
-        try:
-            # If result is already in correct format, validate and return
-            if isinstance(result, dict) and 'response' in result:
-                return {
-                    'response': str(result.get('response', '')),
-                    'confidence': float(result.get('confidence', 0.7)),
-                    'reasoning_trace': result.get('reasoning_trace', {}),
-                    'mode': result.get('mode', engine_used),
-                    'engine_used': engine_used
-                }
-            
-            # Handle string results (direct answers)
-            if isinstance(result, str):
-                return {
-                    'response': result,
-                    'confidence': 0.75,
-                    'reasoning_trace': {
-                        'engine': engine_used,
-                        'query': query,
-                        'result_type': 'string'
-                    },
-                    'mode': engine_used,
-                    'engine_used': engine_used
-                }
-            
-            # Handle complex result objects (extract relevant fields)
-            response_text = str(result)
-            if hasattr(result, 'result'):
-                response_text = str(result.result)
-            elif hasattr(result, 'answer'):
-                response_text = str(result.answer)
-            elif hasattr(result, 'output'):
-                response_text = str(result.output)
-            
-            confidence = 0.70  # Default confidence
-            if hasattr(result, 'confidence'):
-                confidence = float(result.confidence)
-            elif hasattr(result, 'certainty'):
-                confidence = float(result.certainty)
-            
-            reasoning_trace = {'engine': engine_used, 'query': query}
-            if hasattr(result, 'trace'):
-                reasoning_trace.update(result.trace)
-            elif hasattr(result, 'steps'):
-                reasoning_trace['steps'] = result.steps
-            
-            return {
-                'response': response_text,
-                'confidence': confidence,
-                'reasoning_trace': reasoning_trace,
-                'mode': engine_used,
-                'engine_used': engine_used
-            }
-            
-        except Exception as e:
-            logger.error(f"[WorldModel] Result normalization failed: {e}")
-            # Return minimal valid result on error
-            return {
-                'response': f"Reasoning engine {engine_used} completed but result normalization failed: {e}",
-                'confidence': 0.5,
-                'reasoning_trace': {
-                    'engine': engine_used,
-                    'error': str(e),
-                    'raw_result': str(result)[:500]  # Truncate to prevent overflow
-                },
-                'mode': engine_used,
-                'engine_used': engine_used
-            }
+        from .reasoning_normalize import normalize_engine_result as _fn
+        return _fn(self, result, engine_used, query)
     
     def _philosophical_reasoning(self, query: str, **kwargs) -> Dict[str, Any]:
         from .philosophical_reasoning import _philosophical_reasoning
@@ -3556,71 +2107,8 @@ class WorldModel:
         return _fn(self, query)
 
     def get_system_status(self) -> Dict[str, Any]:
-        """Get comprehensive system status"""
-
-        return {
-            "model_version": self.model_version,
-            "observation_count": self.observation_count,
-            "last_observation_time": self.last_observation_time,
-            "intervention_count": self.intervention_manager.intervention_count,
-            "safety_mode": self.safety_mode,
-            "bootstrap_mode": self.bootstrap_mode,
-            "meta_reasoning_enabled": self.meta_reasoning_enabled,
-            "full_meta_reasoning_enabled": getattr(self, "full_meta_reasoning_enabled", False),
-            "self_improvement_enabled": self.self_improvement_enabled,
-            "improvement_running": (
-                self.improvement_running if self.self_improvement_enabled else False
-            ),
-            "components": {
-                "causal_graph": {
-                    "available": CAUSAL_GRAPH_AVAILABLE,
-                    "nodes": (
-                        len(self.causal_graph.nodes) if CAUSAL_GRAPH_AVAILABLE else 0
-                    ),
-                    "edges": (
-                        len(self.causal_graph.edges) if CAUSAL_GRAPH_AVAILABLE else 0
-                    ),
-                },
-                "correlation_tracker": {"available": CorrelationTracker is not None},
-                "intervention_manager": {
-                    "available": INTERVENTION_MANAGER_AVAILABLE,
-                    "queued": len(self.intervention_manager.intervention_queue),
-                },
-                "prediction_engine": {"available": PREDICTION_ENGINE_AVAILABLE},
-                "dynamics_model": {"available": DynamicsModel is not None},
-                "invariant_detector": {"available": INVARIANT_DETECTOR_AVAILABLE},
-                "confidence_calibrator": {"available": CONFIDENCE_CALIBRATOR_AVAILABLE},
-                "router": {"available": ROUTER_AVAILABLE},
-                "meta_reasoning": {
-                    "available": META_REASONING_AVAILABLE,
-                    "enabled": self.meta_reasoning_enabled,
-                    # Note: Include full meta-reasoning component status
-                    "components": {
-                        # Use getattr consistently for all attributes to ensure safety
-                        "motivational_introspection": getattr(self, "motivational_introspection", None) is not None,
-                        "validation_tracker": getattr(self, "validation_tracker", None) is not None,
-                        "transparency_interface": getattr(self, "transparency_interface", None) is not None,
-                        "internal_critic": getattr(self, "internal_critic", None) is not None,
-                        "curiosity_reward_shaper": getattr(self, "curiosity_reward_shaper", None) is not None,
-                        "ethical_boundary_monitor": getattr(self, "ethical_boundary_monitor", None) is not None,
-                        "preference_learner": getattr(self, "preference_learner", None) is not None,
-                        "value_evolution_tracker": getattr(self, "value_evolution_tracker", None) is not None,
-                        "counterfactual_reasoner": getattr(self, "counterfactual_reasoner", None) is not None,
-                        "goal_conflict_detector": getattr(self, "goal_conflict_detector", None) is not None,
-                        "objective_negotiator": getattr(self, "objective_negotiator", None) is not None,
-                    },
-                },
-                "self_improvement": {
-                    "available": META_REASONING_AVAILABLE
-                    and SelfImprovementDrive is not None,
-                    "enabled": getattr(self, "self_improvement_enabled", False),
-                },
-                "safety_validator": {
-                    "available": EnhancedSafetyValidator is not None,
-                    "enabled": getattr(self, "safety_mode", "disabled") == "enabled",
-                },
-            },
-        }
+        from .state_status import get_system_status as _fn
+        return _fn(self)
 
 
 def create_world_model(config: Optional[Dict[str, Any]] = None) -> WorldModel:
